move GroupsToUserPillow and related functionality to its own module
