+++ corehq/apps/users/management/commands/ptop_fast_reindex_groupstousers.py
@@ -1,6 +1,7 @@
 from corehq.apps.groups.models import Group
 from corehq.apps.hqcase.management.commands.ptop_fast_reindexer import ElasticReindexer
-from corehq.pillows.user import GroupToUserPillow, UserPillow
+from corehq.pillows.groups_to_user import GroupToUserPillow
+from corehq.pillows.user import UserPillow
 
 
 class Command(ElasticReindexer):
